- Resolve Flyway rename conflicts (keep develop versions)

- Ensure unique migration versions (V44, V45, V46)

- No duplicate migration versions remain

+++ backend/src/main/java/vn/nguongocso/certification/dto/request/CreateInspectionRequest.java
@@ -20,7 +20,7 @@ public class CreateInspectionRequest {
     private LocalDate sampleSentDate;
 
     @NotEmpty(message = "Vui lòng chọn ít nhất một chỉ tiêu kiểm nghiệm.")
-    private List<Integer> criteriaIds;
+    private List<Long> criteriaIds;
 
     private Boolean confirmDuplicate = false;
 }
